fix(react): revert OdeClientContext splitting
