+++ packages/react/src/core/OdeClientProvider/index.ts
@@ -1,2 +1,1 @@
-export * from "./AppProvider";
 export * from "./OdeClientProvider";
